[internal/driver/mobile] move device keyboard methods into device.go
